Fixed Naming and Removed Defunct Main Files
NOTES:
- DELETED: Top level python files. They were all runnable python files for the entire program, but all except one (run_sim_panda) was running on the current setup. Removed them because they were not using the current setup.
- RENAMED: sim/Sensor/* => sim/Sensors/*. It's not a singular sensor, it is the sensors package.
- MOVED:
- MODIFIED: A lot of files. I needed to fix the import statement for the renamed 'Sensors' package.
- DELETED: sim/Constants.py & sim/print_helpers.py. They're all in the utils/CONSTNATS.py file.'
- MOVED: A bunch of files. Overall file organization is now cleaner.
