+++ config/agent/basic_drone.json
@@ -2,7 +2,7 @@
     "agent": "bullet/bullet3/data/Quadrotor/quadrotor.urdf",
     "sensors": {
         "sensor_1": {
-            "file": "sim/Sensor/basic_camera.json",
+            "file": "sim/Sensors/basic_camera.json",
             "sensor_pos": [
                 0,
                 0,
@@ -15,7 +15,7 @@
             ]
         },
         "sensor_2": {
-            "file": "sim/Sensor/basic_ir_camera.json",
+            "file": "sim/Sensors/basic_ir_camera.json",
             "sensor_pos": [
                 0,
                 0,
