+++ sensor_selection_simulator.py
@@ -1,17 +1,22 @@
+"""_summary_
+    The main overarching file that calls everything else.
+"""
+
 from pathlib import Path
 import argparse
-from sim.World.sim_world import WORLD
+from sim.sim_world import WORLD
 
-def main():
-    """The main function that runs the simulation."""
+def sensor_selection_simulator():
+    """The attachment point to run the program. Takes arguements."""
+    
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "--config",
         help="yaml file to load settings from",
         default="config\\scene\\mountain_range\\scene.yaml",
     )
-    parser.add_argument("--output_dir", help="path to output directory", default="logs")
-    parser.add_argument("--n", help="number of trials", default=1, type=int)
+#    parser.add_argument("--output_dir", help="path to output directory", default="logs") # Unused so far
+#    parser.add_argument("--n", help="number of trials", default=1, type=int) # Unused so far
     args = parser.parse_args()
 
     world_config = Path(args.config)
@@ -22,10 +27,9 @@ def main():
 def run(app):
     """
     Actually runs the app. 
-    TODO: Put this function into its own function that make that file what we call to run the simulation.
     """
     app.run()
 
 
 if __name__ == "__main__":
-    main()
+    sensor_selection_simulator()
