--- sim/Constants.py
@@ -1,3 +1,0 @@
-speed_of_sound = 343.0
-sim_dt = 0.001
-stefan_boltzmann_constant = 5.670374419E-8
