+++ sim/Sensors/basic_camera.json
