+++ sim/Sensors/basic_ir_camera.json
