--- sim/print_helpers.py
@@ -1,7 +1,0 @@
-RESET   = "\033[0m"
-RED     = "\033[31m"
-GREEN   = "\033[32m"
-YELLOW  = "\033[33m"
-BLUE    = "\033[34m"
-MAGENTA = "\033[35m"
-CYAN    = "\033[36m"
