+++ sim/utils/CONSTANTS.py
@@ -1,0 +1,15 @@
+"""_summary_ CONSTANTS File """
+
+# Math
+SPEED_OF_SOUND = 343.0
+SIM_DT = 0.001
+STEFAN_BOLTZMANN_CONSTANT = 5.670374419E-8
+
+# Print Helpers
+RESET   = "\033[0m"
+RED     = "\033[31m"
+GREEN   = "\033[32m"
+YELLOW  = "\033[33m"
+BLUE    = "\033[34m"
+MAGENTA = "\033[35m"
+CYAN    = "\033[36m"
